requirements.txt
tqdm===4.67.1
numpy===2.3.4
PyYAML===6.0.3
pandas===2.3.3
pillow===12.0.0
langchain===1.0.3
langchain-ollama===1.0.0
face_recognition===1.3.0